Move getDirty tests to ModelGetDirtyTest and add missing coverage
- Extract all dirty detection tests from ModelTest into ModelGetDirtyTest
- Add testGetDirtyDateWithoutCast: covers UTCDateTime fields without cast,
  comparing via Document::fromPHP after DateTimeInterface conversion
